MInor changes in PostController

+++ app/Http/Controllers/PostController.php
@@ -10,25 +10,31 @@ class PostController extends Controller
     /**
      * Display a listing of all posts.
      *
-     * @return \Illuminate\View\View
+     * @return \Illuminate\Http\JsonResponse
      */
     public function index()
     {
-        // Retrieve all posts from the database
-        $posts = Post::all();
+        // Retrieve all posts with their related user from the database
+        $posts = Post::with('user')->get();
         
-        // Return the 'posts.index' view with the list of posts
-        return view('posts.index', ['posts' => $posts]);
+        // Return the posts as JSON
+        return response()->json($posts, 200);
     }
 
     /**
      * Store a newly created post in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @return \Illuminate\Http\JsonResponse
      */
     public function store(Request $request)
     {
+        // Validate the request data
+        $request->validate([
+            'title' => 'required|string|max:255',
+            'body' => 'required|string|max:255',
+        ]);
+
         // Get the currently authenticated user
         $user = auth()->user();
         
@@ -45,15 +51,15 @@ class PostController extends Controller
         // Save the post to the database
         $post->save();
         
-        // Redirect to the '/posts' route
-        return redirect('/posts');
+        // Return the created post as JSON
+        return response()->json($post, 201);
     }
 
     /**
      * Display the specified post.
      *
      * @param  int  $id
-     * @return \Illuminate\View\View
+     * @return \Illuminate\Http\JsonResponse
      */
     public function show($id)
     {
@@ -61,15 +67,15 @@ class PostController extends Controller
         // The firstOrFail() method will throw an exception if the post is not found
         $post = Post::with('user')->where('id', $id)->firstOrFail();
         
-        // Return the 'posts.show' view with the post data
-        return view('posts.show', ['post' => $post]);
+        // Return the post as JSON
+        return response()->json($post, 200);
     }
 
     /**
      * Show the form for editing the specified post.
      *
      * @param  int  $id
-     * @return \Illuminate\View\View
+     * @return \Illuminate\Http\JsonResponse
      */
     public function edit($id)
     {
@@ -77,8 +83,8 @@ class PostController extends Controller
         // The firstOrFail() method will throw an exception if the post is not found
         $post = Post::where('id', $id)->firstOrFail();
         
-        // Return the 'posts.edit' view with the post data
-        return view('posts.edit', ['post' => $post]);
+        // Return the post as JSON
+        return response()->json($post, 200);
     }
 
     /**
@@ -86,10 +92,16 @@ class PostController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\RedirectResponse
+     * @return \Illuminate\Http\JsonResponse
      */
     public function update(Request $request, $id)
     {
+        // Validate the request data
+        $request->validate([
+            'title' => 'required|string|max:255',
+            'body' => 'required|string',
+        ]);
+
         // Retrieve the post by its id
         // The firstOrFail() method will throw an exception if the post is not found
         $post = Post::where('id', $id)->firstOrFail();
@@ -101,15 +113,15 @@ class PostController extends Controller
         // Save the updated post to the database
         $post->save();
         
-        // Redirect to the '/posts' route
-        return redirect('/posts');
+        // Return the updated post as JSON
+        return response()->json($post, 200);
     }
 
     /**
      * Remove the specified post from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\RedirectResponse
+     * @return \Illuminate\Http\JsonResponse
      */
     public function destroy($id)
     {
@@ -117,10 +129,15 @@ class PostController extends Controller
         // The firstOrFail() method will throw an exception if the post is not found
         $post = Post::where('id', $id)->firstOrFail();
         
+        // Check if the authenticated user is authorized to delete the post
+        if ($post->user_id !== auth()->id()) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         // Delete the post from the database
         $post->delete();
         
-        // Redirect to the '/posts' route
-        return redirect('/posts');
+        // Return a success message as JSON
+        return response()->json(['message' => 'Post deleted successfully']);
     }
 }
